chore: move MovieService types and schemas to a separate types file
Moved PosterSize, TMDB_MovieEntry and their related Zod schemas from MovieService.ts to a new types.ts file within the same directory. This improves maintainability and readability by separating implementation details from type definitions. Updated imports and barrel exports accordingly.
