+++ src/services/MovieService/index.ts
@@ -1,1 +1,2 @@
 export * from './MovieService';
+export * from './types';
